% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{gompertz}
\alias{gompertz}
\title{Create a vector of mortality rates generated by a heterogeneous gompertz model}
\usage{
gompertz(rho0, m0, beta, age_vec)
}
\arguments{
\item{rho0}{Numeric vector of length n of initial fractions of sub-populations within total population.}

\item{m0}{Numeric vector of length n of initial sub-population mortality rates.}

\item{beta}{Numeric vector of length n of sub-population rates of mortality dynamics.}

\item{age_vec}{Numeric vector of ages from 0 to upper age limit of population, \omega.}
}
\value{
Outputs a numeric vector of mortality rates for the heterogeneous population
of length 0:\omega.
}
\description{
This function produces a vector of mortality rates for a heterogeneous population containing n sub-populations.
}
\examples{
age_vec <- 0:110; rho0 <- c(0.5, 0.5); m0 <- c(0.05, 0.02); beta <- (0.039, 0.039)
mrates <- gompertz(rho0, m0, beta, age_vec)
plot(age_vec, mrates)
}
\keyword{demographics}
\keyword{gompertz}
\keyword{mortality}
\keyword{population}
\keyword{rates}
