% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_vec.R
\name{prod_vec}
\alias{prod_vec}
\title{A function to produce a vector of products}
\usage{
prod_vec(m0, beta, age_vec)
}
\arguments{
\item{m0}{vector of initial mortality rates.}

\item{beta}{vector of ageing rates.}

\item{age_vec}{vector of ages from birth to upper age limit of population $\omega$.}
}
\description{
This function is a helper function which calculates the product to be used in the rho (pop proportion) function.
}
\examples{
prod_vec(m0, beta, age_vec)
}
\keyword{function}
\keyword{helper}
