% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.R
\name{rho}
\alias{rho}
\title{A function to calculate a matrix of population proportions}
\usage{
rho(rho0, m0, beta, age_vec)
}
\arguments{
\item{rho0}{vector of initial population proportions. Length of vector equals number of sub-populations.}

\item{m0}{vector of initial mortality rates.}

\item{beta}{vector of ageing rates.}

\item{age_vec}{vector of ages from birth to upper age limit of population $\omega$}
}
\description{
This function produces a matrix of population proportions to be used in calculating a vector of mortality rates.
}
\examples{
rho()
}
\keyword{population}
\keyword{proportions}
